Add appointment management methods to AppointmentRepository and update AuthService
