+++ backend/src/api/v1/repositories/appointmentRepository.ts
@@ -223,6 +223,17 @@ export interface IAppointmentRepository {
       })
     | null
   >;
+
+  findActiveAppointments(
+    userId: string,
+    tx?: PrismaTransaction
+  ): Promise<Appointment[]>;
+
+  cancelAppointments(
+    appointmentIds: string[],
+    reason: string,
+    tx?: PrismaTransaction
+  ): Promise<void>;
 }
 
 export class AppointmentRepository implements IAppointmentRepository {
@@ -527,6 +538,38 @@ export class AppointmentRepository implements IAppointmentRepository {
         services: { include: { service: true } },
         user: { select: { id: true, name: true, email: true } },
         stylist: { select: { id: true, name: true, email: true } },
+      },
+    });
+  }
+
+  async findActiveAppointments(
+    userId: string,
+    tx?: PrismaTransaction
+  ): Promise<Appointment[]> {
+    const client = this.getClient(tx);
+    return client.appointment.findMany({
+      where: {
+        OR: [{ userId }, { stylistId: userId }],
+        status: { in: ["PENDING", "CONFIRMED"] },
+      },
+      orderBy: { createdAt: "desc" },
+    });
+  }
+
+  async cancelAppointments(
+    appointmentIds: string[],
+    reason: string,
+    tx?: PrismaTransaction
+  ): Promise<void> {
+    const client = this.getClient(tx);
+    await client.appointment.updateMany({
+      where: {
+        id: { in: appointmentIds },
+        status: { in: ["PENDING", "CONFIRMED"] }, // Extra safety check
+      },
+      data: {
+        status: "CANCELLED",
+        notes: reason,
       },
     });
   }
