+++ backend/src/api/v1/repositories/authRepository.ts
@@ -33,17 +33,6 @@ export interface IAuthRepository {
     id: string,
     tx?: PrismaTransaction
   ): Promise<User & { role: { name: string } }>;
-
-  findActiveAppointments(
-    userId: string,
-    tx?: PrismaTransaction
-  ): Promise<Appointment[]>;
-
-  cancelAppointments(
-    appointmentIds: string[],
-    reason: string,
-    tx?: PrismaTransaction
-  ): Promise<void>;
 }
 
 export class AuthRepository implements IAuthRepository {
@@ -117,38 +106,6 @@ export class AuthRepository implements IAuthRepository {
         deactivatedAt: DateTime.now().toJSDate(),
       },
       include: { role: true },
-    });
-  }
-
-  async findActiveAppointments(
-    userId: string,
-    tx?: PrismaTransaction
-  ): Promise<Appointment[]> {
-    const client = this.getClient(tx);
-    return client.appointment.findMany({
-      where: {
-        OR: [{ userId }, { stylistId: userId }],
-        status: { in: ["PENDING", "CONFIRMED"] },
-      },
-      orderBy: { createdAt: "desc" },
-    });
-  }
-
-  async cancelAppointments(
-    appointmentIds: string[],
-    reason: string,
-    tx?: PrismaTransaction
-  ): Promise<void> {
-    const client = this.getClient(tx);
-    await client.appointment.updateMany({
-      where: {
-        id: { in: appointmentIds },
-        status: { in: ["PENDING", "CONFIRMED"] }, // Extra safety check
-      },
-      data: {
-        status: "CANCELLED",
-        notes: reason,
-      },
     });
   }
 }
